+++ src/models/Suggestion.ts
@@ -1,7 +1,8 @@
-import { CompletionItemKind } from "vscode";
+import { CompletionItemKind, SnippetString } from "vscode";
 
 export interface Suggestion { 
   value: string; 
   description: string;
   completion?: CompletionItemKind;
+  text?: string | SnippetString;
 }